refactor: move scheduler creation to core
